Oh baby, I think I fixed it. - It seems that the camera gets fakey "locked" at the size it was set to   when the user grants access to it.   - Listen for status changes, don't set things up until the camera is     unmuted.   - record the size at the time of unmuting.
This all seems to work, but I worry about what would happen if access
to the camera had already been granted when this code runs.
There'd be no way to find out what the size was access was granted.
Fortunately, I believe we only use this code presently in a standalone
appletty-type-thing, so access is always granted in here.


git-svn-id: svn+ssh://src.earth.threerings.net/nenya/trunk@752 ed5b42cb-e716-0410-a449-f6a68f950b19