I'm in ur nenya breaking ur api
We hit a wall in yohoho where we really needed 3 colors on a scene object.
Turns out that colorizations are restricted to 255 even though we stuff it
in a short, and even WE are only up to mid 50s or so with our zations, so
forcing it into a byte hopefully won't pain anybody else that happens to be
using this code.

So we add the ability to have tertiary and quaternary zations on miso objects
while keeping backwards compatability by slotting our two new bytes into 
unused space.


git-svn-id: svn+ssh://src.earth.threerings.net/nenya/trunk@867 ed5b42cb-e716-0410-a449-f6a68f950b19