Fix a Truly Ancient (TM) bug.
The Scene Resolver thread and the dobj/AWT thread both talk to the _blocks map.  And if the resolver tried to get while someone else was doing a put(), that put could decide to resize/rearrange the map.  Which could result in the resolver not finding its block.  Which resulted in annoying missing tile blocks.  So - ConcurrentHashMap it up.

Woo!



git-svn-id: svn+ssh://src.earth.threerings.net/nenya/trunk@894 ed5b42cb-e716-0410-a449-f6a68f950b19