We can't rely on overriding methods in SubtitleChatOverlay to customize them for a target system, because ComicChatOverlay already overrides SubtitleChatOverlay and we wouldn't want to go down the twisty roads of multiple inheritance even if we could.
Instead, I have factored out the customizable stuff into ChatLogic, which has
the dual nice properties of actually working in Java, and of making it very
obvious what is customizable, which was definitely not the case when these
methods were scattered among the four score methods inside SubtitleChatOverlay.
Win win!


git-svn-id: svn+ssh://src.earth.threerings.net/nenya/trunk@1023 ed5b42cb-e716-0410-a449-f6a68f950b19