Switched (partially) to Maven-based dependency resolution. Further work forthcoming to fix up the tests.
Peskily, it looks like I'm going to have to take one for the team and become
the maintainer of a Maven artifact for LWJGL. In spite of repeated requests
over the years for Mavenized artifacts for LWJGL, Mazon resists it (he seems to
think he has to use Maven to *build* LWJGL, which he certainly need not do).
There have been a variety of half-assed attempts to maintain third-party Maven
repositories, all of which have petered out a major version or two ago.


git-svn-id: svn+ssh://src.earth.threerings.net/nenya/trunk@1045 ed5b42cb-e716-0410-a449-f6a68f950b19