Improve Colorization slightly.
Colorizations are not cached, and the root color values are recomputed
every time. We can share those.
Also added an accessor to read the already computed HSV values out
of a Colorization. I suppose I could just make it a public field
or return the array directly. The rootColor variable is already public.
DFU. But now the array will be shared by every instance from the same
classRecord, so it seems more important to make sure nothing goes wrong.